+++ src/main/java/co/edu/konrad/zoologico/persistence/DocumentoPersistence.java
@@ -48,7 +48,7 @@ public class DocumentoPersistence {
      * @param id
      * @return 
      */
-    public DocumentoEntity find(int id){
+    public DocumentoEntity find(Long id){
         return em.find(DocumentoEntity.class, id);
     }
     
@@ -56,7 +56,7 @@ public class DocumentoPersistence {
     * Metodo para borrar un documento segun id
     * @param id
     */
-    public void delete(int id){
+    public void delete(Long id){
        DocumentoEntity entity = em.find(DocumentoEntity.class,id);
         em.remove(entity);
     }
