+++ src/main/java/co/edu/konrad/zoologico/resources/CuidadorResource.java
@@ -45,7 +45,7 @@ public class CuidadorResource {
      */
     @Path("{id: \\d+}")
     @GET
-    public CuidadorDTO obtenerCuidador(@PathParam("id") int Id){
+    public CuidadorDTO obtenerCuidador(@PathParam("id") Long Id){
         CuidadorEntity entity= logic.obtenerCuidadorPorId(Id);
         if(entity == null){
             throw new RuntimeException("El producto solicitado no existe");
@@ -69,7 +69,7 @@ public class CuidadorResource {
      */
     @PUT
     @Path("{id2: \\d+}")
-    public CuidadorDTO actualizarCuidador(CuidadorDTO cuidador,@PathParam("id2") int id){
+    public CuidadorDTO actualizarCuidador(CuidadorDTO cuidador,@PathParam("id2") Long id){
         CuidadorEntity entity = logic.obtenerCuidadorPorId(id);
         if(entity == null){
             throw new RuntimeException("El producto solicitado no existe");
@@ -82,7 +82,7 @@ public class CuidadorResource {
      */
     @Path("{id3: \\d+}")
     @DELETE
-    public void eliminarCuidador(@PathParam("id3") int id){
+    public void eliminarCuidador(@PathParam("id3") Long id){
         CuidadorEntity entity = logic.obtenerCuidadorPorId(id);
         if(entity == null){
             throw new RuntimeException("El producto solicitado no existe");
